Development

+++ src/main/java/com/codecool/crudinventorymanagementapp/employee/RepositoryEmployee.java
@@ -3,6 +3,8 @@ package com.codecool.crudinventorymanagementapp.employee;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface RepositoryEmployee extends CrudRepository<EmployeeModel, Integer> {
+public interface RepositoryEmployee extends CrudRepository<EmployeeModel, Long> {
 }
