+++ src/main/java/com/codecool/crudinventorymanagementapp/inventory/RepositoryInventory.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepositoryInventory extends CrudRepository<InventoryModel, Integer> {
+public interface RepositoryInventory extends CrudRepository<InventoryModel, Long> {
 }
